+++ src/plans/constrained_set_plan.jl
@@ -1,5 +1,5 @@
 """
-    ConstrainedSetObjective{E, MO, PF, IF} <: AbstractManifoldObjective{E}
+    ManifoldConstrainedSetObjective{E, MO, PF, IF} <: AbstractManifoldObjective{E}
 
 Model a constrained objective restricted to a set
 
@@ -18,7 +18,7 @@ where ``$(_tex(:Cal,"C")) ⊂ $(_math(:M))`` is a convex closed subset.
 
 # Constructor
 
-    ConstrainedSetObjective(f, grad_f, project!!; kwargs...)
+    ManifoldConstrainedSetObjective(f, grad_f, project!!; kwargs...)
 
 Generate the constrained objective for a given function `f` its gradient `grad_f` and a projection `project!!` ``$(_tex(:proj))_{$(_tex(:Cal,"C"))}``.
 
@@ -28,7 +28,7 @@ $(_var(:Keyword, :evaluation))
 * `indicator=nothing`: the indicator function ``ι_{$(_tex(:Cal,"C"))}(p)``. If not provided a test, whether the projection yields the same point is performed.
   For the [`InplaceEvaluation`](@ref) this required one allocation.
 """
-struct ConstrainedSetObjective{
+struct ManifoldConstrainedSetObjective{
     E<:AbstractEvaluationType,MO<:AbstractManifoldObjective,PF,IF
 } <: AbstractManifoldObjective{E}
     objective::MO
@@ -36,14 +36,16 @@ struct ConstrainedSetObjective{
     indicator::IF
 end
 
-function ConstrainedSetObjective(
+function ManifoldConstrainedSetObjective(
     f, grad_f, project!!::PF; evaluation::E=AllocatingEvaluation(), indicator=nothing
 ) where {PF,E<:AbstractEvaluationType}
     obj = ManifoldGradientObjective(f, grad_f; evaluation=evaluation)
     if isnothing(indicator)
         if evaluation isa AllocatingEvaluation
             ind(M, p) = (distance(M, p, project!!(M, p)) ≈ 0 ? 0 : Inf)
-            return ConstrainedSetObjective{E,typeof(obj),typeof(project!!),typeof(ind)}(
+            return ManifoldConstrainedSetObjective{
+                E,typeof(obj),typeof(project!!),typeof(ind)
+            }(
                 obj, project!!, ind
             )
         elseif evaluation isa InplaceEvaluation
@@ -52,39 +54,43 @@ function ConstrainedSetObjective(
                 project!!(M, q, p)
                 return distance(M, p, q) ≈ 0 ? 0 : Inf
             end
-            return ConstrainedSetObjective{E,typeof(obj),typeof(project!!),typeof(ind)}(
+            return ManifoldConstrainedSetObjective{
+                E,typeof(obj),typeof(project!!),typeof(ind)
+            }(
                 obj, project!!, ind
             )
         end
     end
-    return ConstrainedSetObjective{E,typeof(obj),typeof(project!!),typeof(indicator)}(
+    return ManifoldConstrainedSetObjective{
+        E,typeof(obj),typeof(project!!),typeof(indicator)
+    }(
         obj, project!!, indicator
     )
 end
 
-function get_cost(M::AbstractManifold, cso::ConstrainedSetObjective, p)
+function get_cost(M::AbstractManifold, cso::ManifoldConstrainedSetObjective, p)
     return get_cost(M, cso.objective, p)
 end
-function get_cost_function(cso::ConstrainedSetObjective, recursive=false)
-    return get_cost_function(cso.objective)
+function get_cost_function(cso::ManifoldConstrainedSetObjective, recursive=false)
+    return get_cost_function(cso.objective, recursive)
 end
-function get_gradient_function(cso::ConstrainedSetObjective, recursive=false)
-    return get_gradient_function(cso.objective)
+function get_gradient_function(cso::ManifoldConstrainedSetObjective, recursive=false)
+    return get_gradient_function(cso.objective, recursive)
 end
-function get_gradient(M::AbstractManifold, cso::ConstrainedSetObjective, p)
+function get_gradient(M::AbstractManifold, cso::ManifoldConstrainedSetObjective, p)
     return get_gradient(M, cso.objective, p)
 end
-function get_gradient!(M::AbstractManifold, X, cso::ConstrainedSetObjective, p)
+function get_gradient!(M::AbstractManifold, X, cso::ManifoldConstrainedSetObjective, p)
     return get_gradient!(M, X, cso.objective, p)
 end
 
 _doc_get_projected_point = """
     get_projected_point(amp::AbstractManoptProblem, p)
     get_projected_point!(amp::AbstractManoptProblem, q, p)
-    get_projected_point(M::AbstractManifold, cso::ConstrainedSetObjective, p)
-    get_projected_point!(M::AbstractManifold, q, cso::ConstrainedSetObjective, p)
+    get_projected_point(M::AbstractManifold, cso::ManifoldConstrainedSetObjective, p)
+    get_projected_point!(M::AbstractManifold, q, cso::ManifoldConstrainedSetObjective, p)
 
-Project `p` with the projection that is stored within the [`ConstrainedSetObjective`](@ref).
+Project `p` with the projection that is stored within the [`ManifoldConstrainedSetObjective`](@ref).
 This can be done in-place of `q`.
 """
 
@@ -98,29 +104,29 @@ function get_projected_point!(amp::AbstractManoptProblem, q, p)
 end
 
 @doc "$(_doc_get_projected_point)"
-get_projected_point(M::AbstractManifold, cso::ConstrainedSetObjective, p)
+get_projected_point(M::AbstractManifold, cso::ManifoldConstrainedSetObjective, p)
 function get_projected_point(
-    M::AbstractManifold, cso::ConstrainedSetObjective{AllocatingEvaluation}, p
+    M::AbstractManifold, cso::ManifoldConstrainedSetObjective{AllocatingEvaluation}, p
 )
     return cso.project!!(M, p)
 end
 function get_projected_point(
-    M::AbstractManifold, cso::ConstrainedSetObjective{InplaceEvaluation}, p
+    M::AbstractManifold, cso::ManifoldConstrainedSetObjective{InplaceEvaluation}, p
 )
     q = copy(M, p)
     cso.project!!(M, q, p)
     return q
 end
 @doc "$(_doc_get_projected_point)"
-get_projected_point!(M::AbstractManifold, q, cso::ConstrainedSetObjective, p)
+get_projected_point!(M::AbstractManifold, q, cso::ManifoldConstrainedSetObjective, p)
 function get_projected_point!(
-    M::AbstractManifold, q, cso::ConstrainedSetObjective{AllocatingEvaluation}, p
+    M::AbstractManifold, q, cso::ManifoldConstrainedSetObjective{AllocatingEvaluation}, p
 )
     copyto!(M, q, cso.project!!(M, p))
     return q
 end
 function get_projected_point!(
-    M::AbstractManifold, q, cso::ConstrainedSetObjective{InplaceEvaluation}, p
+    M::AbstractManifold, q, cso::ManifoldConstrainedSetObjective{InplaceEvaluation}, p
 )
     cso.project!!(M, q, p)
     return q